+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Изменено
 
 - Connect relationship fix [@fringer2423](https://github.com/fringer2423)
+- Review serializer (add answer field)
 
 
 ## [0.0.10] - 2022-10-18